Bug fix: Apply `metric_time` filters to time spine when needed (#1455)
There is currently a bug where, if you query a join_to_timespine metric
with a metric_time filter that is not applied in the group by, you will
end up with more rows than expected.
To fix that, we need to apply any metric_time filters to the time spine
table before joining that table to the aggregated measure. This needs to
happen before the join instead of after because after the join we may
not have access to metric_time at the grain needed for the filter. This
implements that behavior.