Require semantic_model context when getting time dimension's defined granularity
And a lot of downstream changes as a result. Previously, we were getting a time dimension's defined granularity from just its element_name. This is not sufficient information, since there might be multiple dimensions with the same name across semantic models, all with potentially different granularities. This updates those methods to take in a SemanticModelElementReference, which contains both the time dimension's element_name and its semantic model name. This caused a lot of downsream impact, since we needed to get the semantic model context wherever those methods were called. I also did some type cleanup along the way to try to simplify the code with this change in mind.