Update aggregated conversion node to only have queried linkable specs (#1381)
## Context
When you try to query a conversion metric by supplying a filter with a
dimension that exists in the base measure's semantic model without
providing it in the group by, the rendering fails. This is because with
the logic in the conversion metric builder, it ended up "requiring" all
the linkable specs from the query (group by + filter), so the group by
provided in the filter gets added to the resulting `SELECT` even though
it was not supplied in the group by. This caused an error with the join
as it didn't expect that additional element. This PR updates it so that
instead of using all the required linkable specs, we should only use the
queried linkable specs when building out the conversion aggregate node.

Resolves #1210 
Resolves SL-2777