Improve performance of `MetricTimeQueryValidationRule` (#1481)
This PR improves the performance of `MetricTimeQueryValidationRule`. One
of the reasons for the slow performance is that it currently makes calls
to `get_joinable_scd_specs_for_metric`, which can be slow with large
manifests. Instead, this runs checks based on the resolved
group-by-items, which is a much smaller set.