Remove unnecessary logic for retrieving agg time dimension for conversion metrics (#1483)
## Context

For conversion metrics, we use the agg time dimesnion of each measure to
do the conversion window calculation. Since we have `metric_time` which
is already referencing the agg time dimension of a given semantic model,
we don't need to use the actual agg time dimension reference. This
removes the need for additional logic in the code as well as duplication
in the rendered SQL.

**Note:** This should not affect any changes in the query result and
should be just removing unnecessary duplication in the rendered SQL

### Example
Currently, the rendered SQL
```sql
...
SELECT
          , subq_25.ds__day AS ds__day <-- dupe
          , subq_25.ds__day AS metric_time__day  <-- dupe
          , subq_25.user AS user
          , subq_25.visits AS visits
        FROM (
          -- Read Elements From Semantic Model 'visits_source'
          SELECT
            1 AS visits
...
```
Now,
```sql
...
SELECT
          subq_25.ds__day AS metric_time__day
          , subq_25.user AS user
          , subq_25.visits AS visits
        FROM (
          -- Read Elements From Semantic Model 'visits_source'
          SELECT
            1 AS visits
...
```