/* PR_START p--cte 06 */ Split column pruner into two phases.
Currently, the column pruner checks the columns that are needed in each `SELECT`
statement and generates the pruned SQL in a single pass. For better readability
and easier modification, this splits the column pruner into two phases.

First, the SQL nodes are traversed to figure out which columns are required and
which can be pruned. Then, the SQL nodes are rewritten with the pruned columns.