Bug fix: Generate correct instances for `JoinOnEntitiesNode` (#1499)
There is an internal bug in the dataflow to SQL logic for the
`JoinOnEntitiesNode` that does not appear to impact production, but does
impact work I'm doing up the stack. The instances generated in the node
contained column associations that were not correct. They referenced the
old column associations (from the parent dataset), which did not include
the entity links that were added in this node. These incorrect column
associations did not appear to be used for anything, and the select
columns were generated from the specs instead (which is correct), so
that's why this bug flew under the radar.

This PR removes the logic that was causing the bug (the
`AddLinkToLinkableElements` instance converter). Instead, it adds logic
to loop through the parent instances and build new instances & select
columns simultaneously.

This change left a lot of boilerplate & duplicate code in the function,
so I moved a lot of that out to helpers for the instances and specs.

This also adds a lot of simplification for the `JoinOnEntitiesNode`.
It's one of our oldest nodes and was quite verbose.

I recommend reviewing by commit. The snapshot changes are primarily
changes to the column order and some node identifiers. There should be
no actual behavior changes here.