Remove unnecessary FilterElementsNodes
Measure aggregation logic applys several FilterElementsNodes that do not actually impact the outcome of optimized queries. Me and Will are making an effort to align the logic in the dataflow plan builder as much as possible. In order to align measure aggregation logic with other logic in the dataflow plan builder, we need to remove these nodes.