Remove unnecessary `FilterElementsNodes` before measure aggregation (#1506)
We are working to dedupe logic in the dataflow plan builder in order to
reduce bugs. There are currently three different paths the dataflow plan
may take, each of which duplicates much of the dataflow plan building
logic:
- measure aggregation
- secondary aggregation for conversion metrics
- no-metric queries
And I am in the process of adding one more path related to time spine
nodes.

The differences in these paths is frequently the source of bugs that
could be easily prevented. For example, forgetting to include a where
filter spec in the original source node is a bug that frequently pops up
for new paths in the dataflow plan.

Further up the stack, I'm going to dedupe a lot of the dataflow plan
logic in order to prevent these bugs from popping up in the future.
Specifically, I'll move all the logic after building a source node
recipe and before applying aggregations to a helper function that can be
reused in all paths (current or future). That will help to ensure that
applying joins, joining custom granularities, applying where filters,
and applying time constraints, always uses the correct logic.

Measure aggregation logic applies several `FilterElementsNodes` that do
not actually impact the outcome of optimized queries. Other paths do not
have these unnecessary nodes. Removing these nodes is a step needed to
ensure measures use the same logic as other dataflow plans, which will
then allow us to dedupe that logic.