Support expectations in snapshot files (#1513)
To make snapshot files easier to review, this PR adds an
`expectation_description` field to the headers. This way, reviewers can
read the description to check that the snapshot contains the expected
result. The field was named `expectation_description` instead of
`expectation` to make it more clear that it's supposed to be
human-readable text vs. a specific value in the snapshot.