Add a method to figure out common nodes in a dataflow plan (#1520)
This PR adds a method to figure out nodes in a dataflow plan that appear
more than once. i.e. a node that is the parent of multiple nodes. These
common nodes indicate operations where a computation is reused, e.g. a
metric that is used in the computation of multiple derived metrics in a
query. These nodes will be later used to generate CTEs.