Support generation of CTEs in `DataflowToSqlQueryPlanConverter` (#1521)
This PR:
* Adds `SqlGenerationOptionSet` to encapsulate the options for how SQL
should be generated from the dataflow plan.
* Adds `O5` level that uses all previous optimizers and generates CTEs.
* Updates `DataflowToSqlQueryPlanConverter` to use
`SqlGenerationOptionSet`. When the `allow_cte` option it set, converts
the common nodes (as implemented in
https://github.com/dbt-labs/metricflow/pull/1520) in a dataflow plan to
map to a CTE instead of a subquery.

Since CTEs are not generated by default, the generated SQL is the same
for test cases and there are no snapshot changes (aside from the ones
that specifically test this feature).