Consolidate logic around choosing appropriate time spine source
This will change later when we support multiple time spine nodes per query. For now, move the error to the core function so that we don't need to do error handling everywhere this gets used (which will be several places further up the stack). Also adds a helper to improve readability.