`JoinToTimeSpineNode` bug fix (#1541)
We had previously made a product decision about the behavior of the
`JoinToTimeSpineNode` that we later decided was not correct. This is
specifically related to what happens if you query with both
`metric_time` and the `agg_time_dimension` (e.g., `booking__ds`). If you
query with just one or the other, we would use the time spine to fulfill
that dimension. But if you queried with both, there was different
behavior.

Previous behavior when querying with both:
- Treat `metric_time` as the time spine dimension and fulfill those
values with a time spine column
- Treat `agg_time_dimension` as any other dimension, and fulfill those
values from the semantic model source table

This behavior felt inconsistent, since you might get different values
when querying the `agg_time_dimension` depending on if you included
`metric_time` in the query or not. This could be especially confusing if
you included `agg_time_dimension` only in the group by, and
`metric_time` in the where filter.

New behavior when querying with both:
- Treat `metric_time` and `agg_time_dimension` the same. Fulfill both
dimensions with values from the time spine.

PR notes:
- I recommend reviewing by commit.
- I tried to isolate the changes related to the bug fix, but it was not
as easy as I'd hoped. Hopefully it isn't too difficult to understand the
changes, but let me know if you want to sync to discuss.
- There are a bunch of SQL changes here, but only a few are in optimized
snapshots. The optimized ones more clearly demonstrate the actual
behavior change.