Bug fix: apply time constraints after time offsets (#1544)
Time constraints were being applied before time offsets in some
scenarios. This is incorrect because the values will change during the
offset. This fixes that. Customers likely never saw this bug because
time constraints are only available to open source users, and this may
never have been released to them.