Memoize results in `SourceScanOptimizer` for better CTE generation (#1563)
Generation of CTEs is determined by identical nodes in the dataflow
plan. Previously, the `SourceScanOptimizer` would create new nodes
whenever it was able to optimize a branch. In addition, it would create
new nodes in some cases where it was not necessary.

This PR updates `SourceScanOptimizer` to memoize results so that
identical nodes are used in cases where the dataflow branch is the same.

This results in snapshot changes for tests where the optimizer
previously prevented CTEs from being used - please view by commit.