Include Time Spine Nodes in Dataflow Plan (#1548)
Previously, time spine nodes were built on the fly in the dataflow to
SQL logic. This came with several limitations, including:
- We can't apply where constraints and time constraints using the
standard dataflow plan logic. This limitation was the main driver for
this PR stack.
- We can't track time spines as source nodes. We might want to add these
to the dbt DAG at some point.

This PR adds logic to build time spine nodes in the dataflow plan,
instead.