Add improved test for the SQL optimization level in the request (#1566)
This PR adds an improved test to show that the SQL optimization level in
the request is handled correctly in the MF engine API. The previous test
made a more general check, but a snapshot helps to check for specific
features. Considering adding an assertion for the snapshot as well.