Create `SqlAddTimeExpression` (#1562)
Adds a new SQL expression that adds a time interval to a date/time
expression. This differs from the existing time delta expression because
it 1) adds instead of subtracts, and 2) accepts a SQL expression for the
count instead of an integer. This will be used to build the SQL for
custom granularity offset windows.