Add SQL exprs needed for custom offset window (#1575)
This includes a `CASE` expression, an integer expression, an arithmetic
expression, and some updates to the add time expression & window
function expressions. These are all needed to build the following SQL
column:
```
CASE
  WHEN ds__martian_day__first_value__offset + INTERVAL (ds__day__row_number - 1) day <= ds__martian_day__last_value__offset
    THEN ds__martian_day__first_value__offset + INTERVAL (ds__day__row_number - 1) day
  ELSE ds__martian_day__last_value__offset
  END AS metric_time__day
```

Note that the first commit is moving the `sql_exprs` file into
`metricflow-semantics`, which is needed for a commit further up the
stack.