Add support for querying with metric alias (#1573)
## Summary

This PR introduces support for querying with aliased metrics. This is
done by adding an `AliasSpecsNode` after the order by step in
`DataflowPlanBuilder`. To make the alias information reach that point, I
also had to modify some of the query interface classes to allow aliases
in metrics.

You can review commit by commit.

---------

Co-authored-by: Courtney Holcomb <courtneyeholcomb@gmail.com>
Co-authored-by: Will Deng <william.deng14@gmail.com>