Bug fix: don't erase order by in SqlRewritingSubQueryReducer
Previously, when applying the SqlRewritingSubQueryReducer, we would allow the child order bys to overwrite the parent order bys. This was never a problem because order bys were only applied in the outermost SQL query. With the recent change to allow metric aliases, order bys may now be in a parent node, and therefore were being erased. This fixes that issue.