Restore type checking for functions decorated with `@lru_cache` (#1596)
There is an issue with mypy when using the `@lru_cache` decorator that
causes function parameters to not be type-checked. For the most part,
we're migrating MF to use our custom `LruCache` class, but for those
cases where we still use the decorator (or until we migrate all of them
to the new class), this ensures they will be type checked.