For `SqlColumnPrunerOptimizer` - add support for CTEs in sub-queries (#1613)
Previously, the `SqlColumnPrunerOptimizer` only accounted for CTEs
defined at the top level as in current use, the
`DataflowToSqlPlanConverter` only generated SQL plans in that way. To
allow for generation of SQL plans with CTEs defined in sub-queries (at
any depth), this PR updates `SqlColumnPrunerOptimizer` to keep track of
the CTE alias mapping / CTE context when traversing the SQL plan.