Build DataflowPlan for custom offset window (#1584)
This is the dataflow plan that will be used if the custom grain is
queried with any grains that aren't the same as the grain used in the
offset window. There will be a simpler dataflow plan used for queries
using only the same grain as what's used in the offset window.