Fix `SelectStatementNode` Duplication Issue in `DataflowNodeToSqlCteVisitor` (#1618)
This PR fixes an issue where a `SelectStatementNode` was duplicated /
referenced multiple times when generating a `SqlPlan` with CTEs. This
was due to a cache in `DataflowNodeToSqlCteVisitor`, so this updates the
cache so that different elements are stored. The new elements can be
used to create a new `SelectStatementNode` on a cache hit.

In cases where a `SelectStatementNode` is referenced multiple times, the
column pruner will be more conservative and not prune as many columns as
possible. With the fix, the snapshot changes reflect additional columns
that can be pruned.