Run the Github cache workflow using the `main` branch (#1636)
I have been puzzled as to why my PRs sometimes do not get a cache hit in
the test Github workflows. Without caching, even the linting check takes
minutes to run because it needs to install the dependencies. Based on
the docs, it seems that the cache might need to be run on `main`:


https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/caching-dependencies-to-speed-up-workflows#restrictions-for-accessing-a-cache

```
Workflow runs cannot restore caches created for child branches or sibling branches. 
For example, a cache created for the child feature-b branch would not be accessible 
to a workflow run triggered on the parent main branch. 
```

So this PR is trying out an action that will run on a push to main to
populate the cache.