Release `0.207.1` (#1615)
Resolves #1589
Resolves #1632 

* The previous release `0.207.0` was not compatible with `dbt-core=1.9`
as the requirement for `dbt-semantic-interfaces` in both packages was
different.
* To address this, I've cut a branch `release-0.207.1` from commit
`5bb13511d36c6122527a1bd12de9d6e0707a9ddc`, which seems to be the last
version before the dependency requirement for `dbt-semantic-interfaces`
changed in `metricflow`.
* The requirement for `dbt-semantic-interfaces` was previously pinned to
`0.7.2` - I'm looking for context on the strict pin.
* For now, I've updated the requirement to
`dbt-semantic-interfaces>=0.7.4,<0.8`. `0.7.4` was the lowest version
compatible with `dbt-core=1.9`.