Fix off-by-one issue in pretty-formatter (#1665)
The pretty-format methods have a `max_line_length` that is used to keep
the formatted output within the given number of columns when possible.
The argument is supposed to be >0, but in some cases, `max(0, ...)` was
used, resulting in errors when the limit was hit. This PR updates those
calls to `max(1, ...)`.

This PR includes some commits to better organize the tests - please view
by commit.