Update handling for Pydantic objects in `mf_pformat` (#1666)
This PR updates the handling for Pydantic objects in `mf_pformat` to
check if the `dict()` methods exists instead of the
`isinstance(BaseModel)` check. The `BaseModel` is actually from the
`dsi_shim` and in some cases with different Pydantic versions, it's
possible that the check does not pass. Since it's not foolproof, the
formatter falls back to `pprint` if there is an error.