Allow for lazy-evaluation of arguments in `LazyFormat` (#1669)
When creating a `LazyFormat` object, the keyword argument values may be
constructed from expensive methods. Similar to allowing a function to be
passed in for the `message` argument, this PR enables methods to be
passed in for the keyword argument values so that they can be lazily
evaluated.

There were some logging calls that were found in performance profiling
to be expensive because lazy-evaluation was not done, so this PR updates
a few of those call sites as well. An alternative approach would have
been to implement `MetricFlowPrettyFormattable` for those objects.