Update Github actions to use the default Python version (#1671)
I found there were still more actions using Python 3.8. We should be
using Python 3.9, and using the older version affects how actions are
cached. To simplify management, this removes places that specify a
Python version so that the default version (3.9) is used.

There were places where Python 3.12 was specified, but it was unclear
why that was the case. Changed those to use the default version as well.