Consolidate loops in `SqlMapRequiredColumnAliasesVisitor` (#1678)
This PR consolidates a couple of loops in
`SqlMapRequiredColumnAliasesVisitor` and replaces `.add_alias()` calls
with `.add_aliases()` to reduce nesting. There are no logic changes in
this PR, and this is mainly to make
https://github.com/dbt-labs/metricflow/pull/1679 a simple change given
the complexity in that class.

There is some additional consolidation that can be done with other
loops, but leaving those out of scope.