Fix bug with missed columns in the column pruner (#1679)
This PR fixes a bug where the column pruner does not correctly keep
track of columns that are used in `SqlStringExpression` and
`SqlColumnAliasReferenceExpression`. Those expressions are different
from other expressions as they do not have a table alias, and there is
existing case handling for those expressions. The bug is that the
handler did not propagate the knowledge that columns from those
expressions is required to the available CTEs.

This bug currently does not affect queries due to the way that we wrap
access to CTEs in a `SELECT`, but fixing this as it affects later work.