Release `0.207.2` (#1661)
The `0.207.2` release adds a number of commits from `main` that were
previously excluded in `0.207.1` due to the breaking
`dbt-semantic-interfaces` version change. The goal was to include many
commits so that more fixes are included, and so that future backports
would be easier to do. The backported commits can be viewed by comparing the [p--release-0.207.2-updates--01](https://github.com/dbt-labs/metricflow/tree/p--release-0.207.2-updates--01)
branch and the
[release-0.027.2](https://github.com/dbt-labs/metricflow/tree/release-0.207.2)
branch. I left those cherry-picks in a branch since the PR would not
load reasonably.

More commits could be included, but there are more complex conflicts to
resolve.