Update tutorial project to use `ref()` (#1691)
This PR updates the tutorial project to use `ref()` instead of
`source()` in the models. By using `ref()`, `dbt build` will be able to
seed appropriately by determining dependencies, allowing the removal of
the `dbt seed` step.

For context, updates to the tutorial project are helpful for later
changes to the CLI tests / CLI test framework.