Avoid reinitialization of `CLIConfiguration` (#1697)
Currently, the `CLIConfiguration` object is initialized / reloaded at
the start of a CLI command. To aid test setup, this PR updates CLI
commands to only initialize the configuration if it has not already been
done. This allows for a prebuilt configuration to be passed into the CLI
when running tests and also allows for more straightforward invocation
of a CLI command from another CLI command.