Add index to store group-by-item metadata (#1714)
This PR adds `LinkableSpecIndex` that is an index object that can be
stored and used to initialize `MetricflowEngine` more quickly. This is
mainly a shuffling of some of the code in `ValidLinkableSpecResolver` to
`LinkableSpecIndexBuilder`, but it can be hard to see from the commits
what is old code vs. new code. In addition, there is some more
consolidation that can be done between the two classes.

There should be no logical / semantic changes from this update, so the
existing tests are used to validate correctness (e.g.
`test_linkable_spec_resolver.py`).