Provide context in `MetricFlowPrettyFormattable.pretty_format()` (#1733)
This PR updates the signature of
`MetricFlowPrettyFormattable.pretty_format()` to include a format
context object. This allows implementing classes to format itself with
more flexibility and with more adherence to the format options. One
issue with the previous approach was that custom implementations
couldn't properly respect the overall line length limit when formatted
as a part of a nested object.