Generate a `SELECT` statement for sink nodes in a dataflow plan (#1743)
This PR updates the generation of SQL from the sink nodes in a dataflow
plan (`WriteTo*...` nodes) to generate a pass-though `SELECT`. The sink
nodes in a query request were mostly symbolic and it previously retuned
the SQL from the parent node. However, the sink nodes can be used to
generate a `SELECT` statement that reorders columns. As the final SQL
will be simplified by the optimizers, this update does not change the
complexity of the final query.