Update engine interface to order output columns by query inputs (#1746)
This PR adds support for a flag in the API that changes the columns in
the output the follow the order of query inputs. Since the API does not
specify the ordering between group-by items and metrics, columns for
group-by items are chosen to come before the columns for metrics.