Use `output` from the `click` invocation to check test results (#1749)
There were some recent CI test failures that may be the result of
inconsistent or updated behavior from the `Click` dependency. The output
of CLI commands was previously checked with the `stdout` field, but that
field is now returning an empty value in the error case. This PR updates
the test framework to use the `output` field that combines both `stdout`
and `stderr`.