Replace `Singleton` with `@singleton_dataclass()` (#1761)
This PR replaces the existing `Singleton` base class with a
`singleton_dataclass` decorator to make it easier to use. I would have
liked to avoid the use of a custom metaclass, but I wasn't able to find
a better approach. If there's a widely-used library implementation found
later, we can swap it in.