Reclassify user related errors (#1760)
## Context

We throw a bunch of different exceptions for different scenarios. In
attempts to categorize the errors appropriately, we will introduce a
`InformativeError` which denotes that it isn't a system error from MF,
but an error that falls under the bucket of things that a user can tweak
in their input (either in their manifest config or query formulation) or
it's just not supported.

This isn't a complete solution, but there will be further efforts to
classify more and make all the error handling be more verbose.