Metadata query updates (#1771)
Some updates to metadata queries exposed on MFEngine. These will be used
in the API. We have built some of these parameters in the API already,
but putting them in MF will make the queries more efficient (we won't
have to loop through the metadata as many times to get the same result).
New features:
- Search and pagination on list metrics, dimensions, entities, and saved
queries
- A new endpoint to list all group bys (both dimensions and entities).
There is some refactoring in this PR as well to dedupe logic needed in
this endpoint
- Sorting by default: some of these endpoints were doing this already,
so this creates consistency

Note about the `list_group_bys` endpoint: it currently does not expose
entities unless metrics are included. This is intentional due to some
limitations around our current support for querying entities without
metrics, and will be resolved in an upcoming PR once the limitations are
fixed. Discussion of those limitations can be found
[here](https://dbt-labs.slack.com/archives/C047UU1H7R9/p1750717779678569).