Allow including derived time grains in list_group_bys (#1775)
1. Allow including derived time grains in `list_group_bys` - This is
used for a pretty niche feature in our API that allows you to list all
group by names for a given metric. See
[here](https://github.com/dbt-labs/metricflow-server/pull/1498) for how
it's used.
2. Dedupe group bys exposed in `MFEngine`. There might be duplicates
here due to differences between the `LinkableDimension/Entity` and
`Dimension/Entity` objects. Ultimately this is because of different join
paths used to reach the group by from different metrics.