Support for saved query order by & limit (#1759)
Actually apply the `order_by` and `limit` set in a saved query. This was
previously implemented in the YAML spec, but never implemented in
MetricFlow.
If the user queries the saved query with a different `order_by` /
`limit` in the query params, that will override what's set in the saved
query.

Note that this required supporting metric object syntax in saved
queries, as well, in order to apply the `.descending(..)` syntax.