Update the snapshot format for `LinkableElementSet` (#1784)
To enable use of the semantic graph with a control flag, the MF engine
needs a common way of resolving group-by items for a query. To do this,
later PRs will change `LinkableElementSet` to an interface with
different implementations. To keep test snapshots consistent, this PR
updates the format of the snapshot so that the outputs are the same for
both implementations.