Add interfaces for `LinkableElementSet` / `LinkableSpecResolver` (#1785)
As hinted in https://github.com/dbt-labs/metricflow/pull/1784, the
migration to the semantic graph involves adding interfaces for
`LinkableElementSet` / `LinkableSpecResolver` and then providing
different implementations (one using the existing classes, another using
the semantic graph). This PR:

* Adds `AnnotatedSpec` that has similar properties as the `Linkable*`
classes.
* Adds the `BaseLinkableElementSet` interface and makes
`LinkableElementSet` implement that interface.
* Renames `LinkableSpecResolver` to `LegacyLinkableSpecResolver` and
makes `LinkableSpecResolver` an interface.
*  Migrates existing code to use those interfaces.

The lack of parallelism in how the interfaces are named is confusing,
but the rename to address that ended up making too many conflicts, so
we'll have to do the rename later.

These intermediate / temporary classes will need to be refined
post-migration. Please view by commit.