Add a helper for writing `timeit`-based performance tests (#1788)
This PR adds a helper that makes it easier to write tests that compare
performance of function calls using `timeit`. The issue with `timeit` is
that the arguments are strings of code snippets. Those strings are
annoying to write as the IDE inspections don't work / `mypy` does not do
type checks. In addition, all required imports must be included in the
code snippets.

The helper provides a closure that can be written normally and is
instantiated in the `timeit` statement. Please view by commit as there
were some renames that are included.

The helper is a WIP so this PR does not migrate existing performance
tests.