Add fast lookup for semantic-manifest objects (#1789)
From the docstring:

These streamlined / minimal lookups were added to use for initializing
the semantic graph as the current lookup
classes have significant initialization times relative to initialization
time of the semantic-graph-based resolver.
Using an example high-complexity semantic manifest, initialization of
the current lookups took ~0.2s.
Initialization of the semantic-graph resolver with the same manifest was
< 1s, so using the current lookups would
have resulted in a significant relative increase.

Initialization of this lookup and an example call is 100x faster (see
`test_model_lookup_performance`).

Future work includes further refinement of these classes and replacing
existing `*Lookup` classes with the ones in
this module / other implementations.