Migrate back to `Singleton` for simplicity / performance (#1793)
The `@singleton_dataclass` decorator was added in
https://github.com/dbt-labs/metricflow/pull/1761 to make singletons
easier to retrieve, and there was a small performance penalty with using
them due to the custom metaclass.

The custom metaclass was a little too off-the-beaten-path and for the
semantic-graph implementation, the performance difference made a
not-insignificant impact, so this reverts back to the old `Singleton`.