Add subgraph generator for metrics (#1802)
This PR adds a subgraph generator to model the relationship between
metrics. There are a few special cases related to labeling the edges
between metrics and their inputs. Normally, the group-by items available
for a metric are an intersection of the available group-by items of the
input metrics or input measures. However, there are a few special cases
(e.g. cumulative metrics) that are more restrictive than this general
rule. These labels will be later used to limit graph traversal.