Switch back to using minimum time grain from models (#1822)
The semantic graph implementation initially used the minimum grain used
in a semantic model as the minimum valid grain (original comment here:
https://github.com/dbt-labs/metricflow/blob/4a2edcfbf272f4cc9aef998fd51475b9d8040aad/metricflow-semantics/metricflow_semantics/experimental/semantic_graph/builder/time_entity_subgraph.py#L43),
but then later was changed to use the minimum grain of the time spine as
it seemed like it was the correct behavior. To aid migration however,
this changes the logic back to the initial implementation.